change in edit functionality to edit once click on the edit button
